Move refcase tests out of test_ensemble_config
